fold Stack into Editor
