+++ packages/slate-prop-types/src/index.js
@@ -10,7 +10,6 @@ import {
   Node,
   Range,
   Schema,
-  Stack,
   Value,
   Text,
 } from 'slate'
@@ -72,7 +71,6 @@ const Types = {
   range: create('Range', v => Range.isRange(v)),
   ranges: create('List<Range>', v => Range.isRangeList(v)),
   schema: create('Schema', v => Schema.isSchema(v)),
-  stack: create('Stack', v => Stack.isStack(v)),
   value: create('Value', v => Value.isValue(v)),
   text: create('Text', v => Text.isText(v)),
   texts: create('List<Text>', v => Text.isTextList(v)),
