--- packages/slate-simulator/Readme.md
@@ -1,1 +1,0 @@
-This package contains a simulator for testing Slate editors and plugins.
