--- packages/slate-simulator/package.json
@@ -1,38 +1,0 @@
-{
-  "name": "slate-simulator",
-  "description": "A simulator for testing Slate editors and plugins.",
-  "version": "0.4.66",
-  "license": "MIT",
-  "repository": "git://github.com/ianstormtaylor/slate.git",
-  "main": "lib/slate-simulator.js",
-  "module": "lib/slate-simulator.es.js",
-  "umd": "dist/slate-simulator.js",
-  "umdMin": "dist/slate-simulator.min.js",
-  "files": [
-    "dist/",
-    "lib/"
-  ],
-  "peerDependencies": {
-    "slate": ">=0.32.0"
-  },
-  "devDependencies": {
-    "slate": "^0.41.2"
-  },
-  "scripts": {
-    "clean": "rm -rf ./dist ./lib ./node_modules"
-  },
-  "umdGlobals": {
-    "slate": "Slate"
-  },
-  "keywords": [
-    "ci",
-    "editor",
-    "simulate",
-    "simulator",
-    "spec",
-    "slate",
-    "test",
-    "testing",
-    "unit"
-  ]
-}
