+++ support/rollup/config.js
@@ -9,7 +9,6 @@ import slateHyperscript from '../../packages/slate-hyperscript/package.json'
 import slatePlainSerializer from '../../packages/slate-plain-serializer/package.json'
 import slatePropTypes from '../../packages/slate-prop-types/package.json'
 import slateReact from '../../packages/slate-react/package.json'
-import slateSimulator from '../../packages/slate-simulator/package.json'
 
 const configurations = [
   ...factory(slate),
@@ -22,7 +21,6 @@ const configurations = [
   ...factory(slatePlainSerializer),
   ...factory(slatePropTypes),
   ...factory(slateReact),
-  ...factory(slateSimulator),
 ]
 
 export default configurations
